go.crypto/ssh/terminal: Make ^D delete current character instead of exit.
Like ^A, ^E, ^U and ^L this is a widespread keyboard shortcut.

The terminal used to return io.EOF immediately, on any ^D keypress.

The terminal now returns io.EOF on ^D only when the input line is empty.

If the cursor is at the end of the input line then ^D now does nothing,
and when the cursor is somewhere within the input line the character
under the cursor is deleted.

Signed-off-by: Isis Lovecruft <isis@torproject.org>